# backend/openapi.yml
openapi: 3.0.3
info:
  title: Notcery API
  version: 1.0.0
  description: |
    Akıllı öğrenme ekosistemi REST API sözleşmesi.

    **Kimlik:** E-posta/şifre ve Google ID Token (`POST /auth/google`).
    **Tercihler:** `locale` (tr, en) ve `theme` (light, dark, system) kullanıcı profilinde.
    **Lokalizasyon:** İstemci `Accept-Language` gönderir; hata mesajları bu dile göre dönebilir.

    Kaynak doğruluk: bu dosya. Backend Faz 1+ implementasyonu buna uyar.
  contact:
    name: Notcery
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.notcery.app/api/v1
    description: Production
...
paths:
...
  /auth/google/link:
    post:
      tags: [Auth]
      summary: Oturum açıkken Google hesabını bağla
      operationId: googleLink
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
      responses:
        '200':
          description: Google hesabı bağlandı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Google hesabı başka kullanıcıya bağlı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Access token yenile
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Yeni token çifti
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token (Authorization header)
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      description: |
        İstenen dil. `tr` veya `en` (BCP 47 kısaltılır).
        Hata mesajları ve opsiyonel sunucu metinleri için.
      schema:
        $ref: '#/components/schemas/Locale'
...
  responses:
    BadRequest:
      description: Doğrulama hatası
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Kimlik doğrulama gerekli veya geçersiz token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Locale:
      type: string
      enum: [tr, en]
      description: Desteklenen arayüz dilleri (MVP)
      example: tr
    Theme:
      type: string
      enum: [light, dark, system]
      description: |
        light / dark — sabit tema.
        system — işletim sistemi tercihini izle (frontend uygular).
      example: system
    AuthProvider:
      type: string
      enum: [email, google]
...
    User:
      type: object
      required:
        - id
        - email
        - locale
        - theme
        - timezone
        - auth_provider
        - onboarding_completed
        - created_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
          nullable: true
        avatar_url:
          type: string
          format: uri
          nullable: true
        locale:
          $ref: '#/components/schemas/Locale'
        theme:
          $ref: '#/components/schemas/Theme'
        timezone:
          type: string
        auth_provider:
          $ref: '#/components/schemas/AuthProvider'
        email_verified:
          type: boolean
          default: false
        onboarding_completed:
          type: boolean
        study_preferences:
          type: object
          additionalProperties: true
        social_accounts:
          type: array
          items:
            $ref: '#/components/schemas/SocialAccountSummary'
          description: Bağlı harici hesaplar (şifre gösterilmez)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SocialAccountSummary:
      type: object
      required: [provider, linked_at]
      properties:
        provider:
          type: string
          enum: [google]
        linked_at:
          type: string
          format: date-time
...
    GoogleAuthRequest:
      type: object
      required: [id_token]
      properties:
        id_token:
          type: string
          description: Google Identity Services credential JWT
        locale:
          $ref: '#/components/schemas/Locale'
          description: İlk kayıtta varsayılan dil (yoksa Accept-Language)
        theme:
          $ref: '#/components/schemas/Theme'
        timezone:
          type: string
          description: IANA timezone; yoksa UTC veya frontend tahmini
    RefreshRequest:
      type: object
      required: [refresh]
      properties:
        refresh:
          type: string
          description: Refresh token
    TokenPair:
      type: object
      required: [access, refresh, access_expires_in]
      properties:
        access:
          type: string
        refresh:
          type: string
        access_expires_in:
          type: integer
          description: Saniye cinsinden access token ömrü
          example: 900
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - ACCOUNT_EXISTS_LINK_REQUIRED
                - GOOGLE_TOKEN_INVALID
                - PLAN_GENERATION_FAILED
            message:
              type: string
              description: İnsan okunur; Accept-Language ile lokalize edilebilir
            details:
              type: object
              additionalProperties: true
...
security:
  - bearerAuth: []